requirements.txt
Django==5.0.1
djangorestframework==3.14.0
requests==2.31.0
python-dotenv==0.19.2